requirements.txt
Flask
python-dotenv
confluent-kafka
prometheus-client
gunicorn